app/requirements.txt
aiohttp==3.9.3
anyio==3.6.2
aiosqlite==0.19.0
apache-libcloud==3.7.0
asyncpg==0.29.0
certifi==2023.5.7
charset-normalizer==3.1.0
datefinder==0.7.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
pillow==10.2.0
pydantic==1.10.8
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.30.0
sniffio==1.3.0
SQLAlchemy==2.0.13
sqlalchemy-file==0.4.0
typing_extensions==4.5.0
urllib3==2.0.2
Faker==24.4.0
flake8==3.9.2
flake8-docstrings==1.6.0